+++ src/dipole/get_tensor.cpp
@@ -46,8 +46,8 @@ namespace dipole{
             in = dipole::internal::rij_tensor_zz;
             break;
       }
-      for(uint64_t lc=0; lc<dipole::internal::cells_num_local_cells; lc++){
-      	for(uint64_t j=0; j<in[lc].size(); j++){ out.push_back( in[lc][j] ); }
+      for(int64_t lc=0; lc<dipole::internal::cells_num_local_cells; lc++){
+      	for(size_t j=0; j<in[lc].size(); j++){ out.push_back( in[lc][j] ); }
       }
       return out;
    }
@@ -76,8 +76,8 @@ namespace dipole{
             in = dipole::internal::rij_tensor_zz;
             break;
       }
-      for(uint64_t lc=0; lc<dipole::internal::cells_num_local_cells; lc++){
-      	for(uint64_t j=0; j<in[lc].size(); j++){ out.push_back( in[lc][j] ); }
+      for(int64_t lc=0; lc<dipole::internal::cells_num_local_cells; lc++){
+      	for(size_t j=0; j<in[lc].size(); j++){ out.push_back( in[lc][j] ); }
       }
       return out;
    }
